Adds endpoints for groups (#43)
* Bump tooling

* Adds groups

* Updates cassettes for accounts

* Finalize groups

* Adds documentation

* Tweak a couple of minor things

* Bump rubocop-mhenrixon minimum version

* Da fuck⁉️

* Prevent install of rspec-nc on linux

* Don't use potentially not used formatters

+++ spec/support/shared_contexts.rb
@@ -3,7 +3,7 @@
 RSpec.shared_context 'with existing account' do
   let!(:account) do
     response = client.create_account(account_params)
-    response.fetch(:account) { raise 'HELL (account creation failed)' }
+    response.fetch(:account) { raise "HELL (account creation failed) #{response}" }
   end
 
   let(:account_id)   { account[:id] }
@@ -17,14 +17,14 @@ RSpec.shared_context 'with existing account' do
   end
 
   after(:each) do
-    client.delete_account(account_id)
+    client.delete_account(account_id) if account_id
   end
 end
 
 RSpec.shared_context 'with existing contact' do
   let!(:contact) do
     response = client.create_contact(contact_params)
-    response.fetch(:contact) { raise 'HELL (contact creation failed)' }
+    response.fetch(:contact) { raise "HELL (contact creation failed) #{response}" }
   end
 
   let(:contact_id)         { contact[:id] }
@@ -42,7 +42,7 @@ RSpec.shared_context 'with existing contact' do
   end
 
   after(:each) do
-    client.delete_contact(contact_id)
+    client.delete_contact(contact_id) if contact_id
   end
 end
 
@@ -52,7 +52,7 @@ RSpec.shared_context 'with existing account_contact' do
 
   let!(:account_contact) do
     response = client.create_account_contact(account_contact_params)
-    response.fetch(:account_contact) { raise 'HELL (account_contact creation failed)' }
+    response.fetch(:account_contact) { raise "HELL (account_contact creation failed) #{response}" }
   end
 
   let(:account_contact_id) { account_contact[:id] }
@@ -66,14 +66,14 @@ RSpec.shared_context 'with existing account_contact' do
   end
 
   after(:each) do
-    client.delete_account_contact(account_contact[:id])
+    client.delete_account_contact(account_contact_id) if account_contact_id
   end
 end
 
 RSpec.shared_context 'with existing address' do
   let!(:address) do
     response = client.create_address(address_params)
-    response.fetch(:address) { raise 'HELL (address creation failed)' }
+    response.fetch(:address) { raise "HELL (address creation failed) #{response}" }
   end
 
   let(:address_id) { address[:id] }
@@ -100,7 +100,30 @@ RSpec.shared_context 'with existing address' do
   end
 
   after(:each) do
-    client.update_address(address[:id], is_default: false)
-    client.delete_address(address[:id])
+    if address_id
+      client.update_address(address_id, is_default: false)
+      client.delete_address(address_id)
+    end
+  end
+end
+
+RSpec.shared_context 'with existing group' do
+  let!(:group) do
+    response = client.create_group(group_params)
+    response.fetch(:group) { raise "HELL (group creation failed) #{response}" }
+  end
+
+  let(:group_id)          { group[:id] }
+  let(:group_title)       { 'Awesome' }
+  let(:group_description) { 'My really awesome group' }
+  let(:group_params) do
+    {
+      title: group_title,
+      descript: group_description
+    }
+  end
+
+  after do
+    client.delete_group(group_id) if group_id
   end
 end
